fix(internal): ignore error for empty folder in config set cmd
When changing params using the `stamusctl config set` command, if a folder is empty we try to delete it.
It leds to an error beeing thrown if the user does not have the permission to delete it.
This error can be ignored as there is no real issue if we cannot delete the folder.