ci: using commitlint and semantic-release
We are using commitlint to lint commit messages from merge requests in the CI. The job in gitlab checks commits included in the merge request and fails if any of them do not follow the conventional commit format.
Semantic-release is used to automatically determine the next version number based on the commit messages and create a release in gitlab.